+++ src/main/java/com/swyp/plogging/backend/post/controller/PostController.java
@@ -3,9 +3,9 @@ package com.swyp.plogging.backend.post.controller;
 import com.swyp.plogging.backend.auth.domain.CustomOAuth2User;
 import com.swyp.plogging.backend.common.dto.ApiPagedResponse;
 import com.swyp.plogging.backend.common.dto.ApiResponse;
-import com.swyp.plogging.backend.controller.dto.CreatePostRequest;
-import com.swyp.plogging.backend.controller.dto.PostDetailResponse;
-import com.swyp.plogging.backend.controller.dto.PostInfoResponse;
+import com.swyp.plogging.backend.post.controller.dto.CreatePostRequest;
+import com.swyp.plogging.backend.post.controller.dto.PostDetailResponse;
+import com.swyp.plogging.backend.post.controller.dto.PostInfoResponse;
 import com.swyp.plogging.backend.post.sevice.ParticipationService;
 import com.swyp.plogging.backend.post.sevice.PostService;
 import org.springframework.data.domain.Page;
