+++ src/main/java/com/swyp/plogging/backend/post/domain/Post.java
@@ -1,7 +1,6 @@
 package com.swyp.plogging.backend.post.domain;
 
-import com.swyp.plogging.backend.controller.dto.PostDetailResponse;
-import com.swyp.plogging.backend.domain.Certification;
+import com.swyp.plogging.backend.post.controller.dto.PostDetailResponse;
 import com.swyp.plogging.backend.domain.base.BaseEntity;
 import com.swyp.plogging.backend.user.domain.AppUser;
 import jakarta.persistence.*;
